prototype/src/frontend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=frontend/requirements.txt frontend/requirements.in
#
click==7.1.2              # via flask
flask==1.1.2              # via -r frontend/requirements.in
grpcio==1.33.2            # via -r frontend/requirements.in
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
prometheus-client==0.9.0  # via -r frontend/requirements.in
protobuf==3.13.0          # via -r frontend/requirements.in
six==1.15.0               # via grpcio, protobuf
werkzeug==1.0.1           # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
